pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "baloise_ai_pylib"
version = "0.0.1"
description = "Balsy Partner System"
license = "Apache 2.0"
authors = ["AI Circle <aicommunity@groups.baloise.com>"]

[tool.poetry.dependencies]
python = "^3"
openai = "^1"
python-dotenv = "^1"
mypy = "^1"
requests ="^2"
types-requests = "^2"